requirements.txt
# VideoGraphRAG End-to-End Pipeline Dependencies
# Python 3.10+

# ── Core & Scientific ────────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
networkx>=3.1

# ── Computer Vision & ML ─────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# ── Retrieval & Graph ────────────────────────────────────────────────────────
neo4j>=5.24.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.2

# ── LLMs & APIs ──────────────────────────────────────────────────────────────
google-genai>=1.16.0
openai>=1.51.0
requests>=2.31.0

# ── Backend & API ────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.8.0

# ── Utilities ────────────────────────────────────────────────────────────────
pyyaml>=6.0
tqdm>=4.65.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
